pyproject.toml
[project]
name = "python-intan"
version = "0.0.3"
description = "Python interface and utilities for working with Intan RHX devices and electrophysiology data"
authors = [
    { name = "Jonathan Shulgach", email = "jshulgac@andrew.cmu.edu" },
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["intan", "emg", "neurotech", "signal processing", "biomedical"]
dependencies = [
    "numpy>=1.26.0",
    "scipy==1.14.1",
    "matplotlib",
    "pandas",
    "PySerial",
    "scikit-learn",
    "tqdm",
    # "nvidia-cuda-toolkit", # This is not a Python package, but a system package
    "tensorflow==2.19.0",
    "seaborn",
	"pyyaml",
    "PyQt5",
    "pyqtgraph",
    "pylsl",
]

[project.urls]
Homepage = "https://github.com/Neuro-Mechatronics-Interfaces/python-intan"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "myst-parser",
]

[tool.setuptools]
packages = ["intan"]

[tool.setuptools.dynamic]
version = {attr = "pyoephys.__version__"}

